End Game Event Script and Noose Death Script + fixes
Created the End Game Event script, but still need to create the new
scene of our "Thanks for Playing" final message. As well as an in-game
"final message" to the player saying only death/darkness awaits those
trapped. Haha

Noose Death Script initially made. Need to fix how the chair tumbles.

Optimized the PillarEvent script a bit to stop some slowdown. Will need
to also optimize the changing/glowing colors of pillarWall so
performance does not drop.

Added all doors needing the keys.

We need to start on the manipulation script and on enemies.

Furniture and content inside the maze can wait until near the end when
we have implemented the heart rate monitor right and working.