Added buttons for the calibration/measurement scene
Added a "Measure" HRV button, a "Reset Rates" button, and a
non-implemented "Start Game" button.

Will have to implement last button still. Also, we will have to convert
the OSC scripts to C# if possible soon. Or make a javascript that sends
the data from one scene to another specifically for these values.