death scripts and static
added a death script to the pillar guardian and made a portion of the
kill script accept a gameobject to be disabled so the guardian does not
keep trying to follow the player. also made the start of a static script
that will increase the static based on the proximity to the guardian.