fixed doors again
deleted a lot of doors and replaced them with functional doors. this is
beyond a joke now. there was also a bunch of bad doors in the final
rooms, so i fixed them up. the random spinning motion that happens to
doors sometimes is caused by rotation, it needs to be -180 or 180
degrees rotation along the axis or it will spaz out