Fixed PlayerChanges.cs. Added EnvironmentChanges.cs that changes the color of the torches in the maze
Initial fix for the PlayerChanges.cs to run better and have better
logic.

Added the EnvironmentChanges.cs thatt only so far changes the color of
the particles and lights of torches. Also plays a sound when it changes
to the scarier color.

Also organized the hierarchy list of gameobjects. As well as a number of
torches having duplicate particles.

WIll look and hopefully get the basics of wall changing down tomorrow
(or continue Sunday evening)