Added wall changing for other walls + thresholds
Added wall changing code fro the walls that are not the same as the
default/main wall we are using. (The T-intersections, mainly in the
cross-section, as well as the door thresholds).

Tested with automatic changing as well as with lightstone. No drops in
normal performance so far.

I think we should use the brick background as a base for the next set of
materials and textures for the change, so we won't have to waste some
time in testing how the textures "fit" each other.

Still need to find a solution for the door top fillers. Maybe a static
gargoyle head like stone wall texture instead?

So far, this code and implementation is good. I want to try something
that may save all the looping through, but I'll do that when we get all
the other features locked and finished.

MORE TESTING! As well with Oculus!