Created Proximity Detector Enlarger code
As title says.

Also tested out the changing wals. Works great. Now all that's left are
the proper textures.

Need to do a balancing and testing for all the aspects of the game. As
well as changing some sounds. If we have time, make a death animation (I
can maybe start that tomorrow).