Various visual additions and some fixes
Added a Floor object that is the same as the ceiling object. Much easier
to change floor textures in this using existing code. Will start on that
tonight.

Walking feels weird with the Oculus. Might need to address that in the
following days.

Fixed various wrong elements in the maze.

Regarding textures: Eyeballs are good. Skulls need changing.

Death animation will really have to be done cause I didn't know I died
in some parts.