Roof and Floor Changing Done + Final Scare Event
Roof and Floor Changing textures implemented. The Floor are all planes,
but if we want to do the terrain idea, we could look at in in the next
few days.

Performance with everything changing isn't too bad. But then again, this
is with the editor (and sometimes both game and scene views are
visible), so the build of the game might work better.

Eyeball in roof and floor with walls looks like I'm in a horror themed
acid trip. lol

Started, and surprisingly half finished the implementation, the final
scare event script. I will still have to change the sprite, audio
design, and possibly do some environment/ambiance change. Will do it in
the next few days.

Start screen also needs to be done.

IMPORTANT: Found a bug in the pillar room. If you continously walk in
the other side/upper right side of the sphere, you will eventually phase
through. I'm not sure how this happens. Do you have any clue Graeme?