Changed bgm and walls and started start screen
Started the start screen of the game. Will be doing some photoshop
textures for it in the next hour

Fixed some bugs and walls that didn't change. Made eyeball the second
texture, while demon faces the last.

started sound design for the last enemy. need better sounds and lines
though.