Start Screen Implemented and Balance Checks
Created (and finished?) start screen.

Change blend speed of walls to 2. I didn't manage to activate the third
texture in some of my tests, so I wonder if we need to add another spook
event in the 4th area? I will think about this on Sunday and Monday.

Almost done. Just need to complete the questions as well