Final Enemy changes + Death animation/indicator
Added voice files for final enemy event. I'm still wondering if we need
to do one more scare event in the end area (or one more enemy).

Made changes to audio volume levels and such.

Created a temp death animation/indicator. It so far does these in order:
-Death Scream
-Disables movement and mouse movement
-Blackouts to darkness/red fog (I'm leaning on red)
-Respawn on checkpoint enabling everything disabled and resetting the
fog to 0

I think I want to change the sprite of the phase through enemy, or at
least make it a darker variation of Mr. Smiles just to have a bit of
consistency in regards to game design.

Questions have also been completed. If Regan is available tomorrow, we
should probably go and ask her if these are appropriate.